[IMP] : - Rename the module to web_widget_color - Remove special type (Color) and use (Char) instead of.
