--- web_color/view/web_color_view.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        <template id="assets_backend" name="web_color assets" inherit_id="web.assets_backend">
-            <xpath expr="." position="inside">
-                <link rel="stylesheet" href="/web_color/static/src/css/widget.css"/>
-                <script type="text/javascript" src="/web_color/static/lib/jscolor/jscolor.js"></script>
-                <script type="text/javascript" src="/web_color/static/src/js/widget.js"></script>
-            </xpath>
-        </template>
-    </data>
-</openerp>
