+++ web_widget_color/doc/picker.png
