+++ web_widget_color/static/lib/jscolor/arrow.gif
