+++ web_widget_color/static/lib/jscolor/demo.html
