+++ web_widget_color/static/lib/jscolor/hv.png
