+++ web_widget_color/static/lib/jscolor/jscolor.js
