+++ web_widget_color/static/src/css/widget.css
