+++ web_widget_color/static/src/xml/widget.xml
